+++ CHANGELOG.md
@@ -13,12 +13,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
 ### Changed
+- MessageAggregator automatically terminates the message domains of the aggregated messages. Can be disabled with optional constructor parameter.
 - **Breaking Change:** Removed the `Predecessors` and `GetPredecessor` members from `Message`. This was a huge unfixable memory leak. Instead of it `MessageDomain` and `MessageCollector` should be used to address scenarios where the predecessor message must be evaluated
 - `MessageAggregator` automatically terminates the message domains of the aggregated messages. Can be disabled with optional constructor parameter
 - `MessageCollector` automatically removes messages from a terminated domain. This was a huge memory leak
 
+### Added
+- `MessageCollector.PushAndExecute` is a replacement for `MessageCollector.FindSetsForDomain`. It adds a message to the collector and waits for a complete set with that message to execute an action
+
 ### Removed
 - **Breaking Change:** Removed useless `MessageDomainsCreatedMessage` and `MessageDomainsTerminatedMessage`
+- **Breaking Change:** Removed the `MessageCollector.FindSetsForDomain` method. Use the new `MessageCollector.PushAndExecute` method instead
 
 ### Fixed
 - `MessageDomain` did not dispose replaced messages. This is now fixed